java: add java:server generator
