+++ generators/java/generators/bootstrap/generator.ts
@@ -73,6 +73,32 @@ export default class BootstrapGenerator extends BaseApplicationGenerator {
     return this.delegateTasksToBlueprint(() => this.configuring);
   }
 
+  get loading() {
+    return this.asLoadingTaskGroup({
+      loadEnvironmentVariables({ application }) {
+        application.packageInfoJavadocs?.push(
+          { packageName: `${application.packageName}.aop.logging`, documentation: 'Logging aspect.' },
+          { packageName: `${application.packageName}.management`, documentation: 'Application management.' },
+          { packageName: `${application.packageName}.repository.rowmapper`, documentation: 'Webflux database column mapper.' },
+          { packageName: `${application.packageName}.security`, documentation: 'Application security utilities.' },
+          { packageName: `${application.packageName}.service.dto`, documentation: 'Data transfer objects for rest mapping.' },
+          { packageName: `${application.packageName}.service.mapper`, documentation: 'Data transfer objects mappers.' },
+          { packageName: `${application.packageName}.web.filter`, documentation: 'Request chain filters.' },
+          { packageName: `${application.packageName}.web.rest.errors`, documentation: 'Rest layer error handling.' },
+          { packageName: `${application.packageName}.web.rest.vm`, documentation: 'Rest layer visual models.' },
+        );
+
+        if (application.defaultPackaging === 'war') {
+          this.log.info(`Using ${application.defaultPackaging} as default packaging`);
+        }
+      },
+    });
+  }
+
+  get [BaseApplicationGenerator.LOADING]() {
+    return this.delegateTasksToBlueprint(() => this.loading);
+  }
+
   get preparing() {
     return this.asPreparingTaskGroup({
       applicationDefaults({ application }) {
